requirements.txt
scipy==0.19.1
tqdm==4.28.1
numpy==1.13.3
pandas==0.22.0
tensorflow==1.10.1
Django==2.1.3
matplotlib==2.1.1
seaborn==0.9.0
Pillow==5.4.1
django-picklefield==2.0
scikit_learn==0.20.3
pydot==1.4.1
h5py==2.9.0
